pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cortex"
version = "0.1.0"
description = "Project memory layer for agentic coding"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "cognee==0.5.5",
  "mcp>=1.20.0,<2.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "uvicorn>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
